# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tarcza Polska API",
        "description": "Backend for the Tarcza Polska civil resilience platform.\n\n* `/api/v1/*` - Citizen mobile app (Flutter). Authenticate with the JWT returned by `POST /api/v1/devices`.\n* `/api/command/*` - Command Center. Authenticate with the JWT returned by `POST /api/command/login`.\n\nAll spatial payloads are GeoJSON (WGS84, `[lng, lat]` order). Every error is an `ErrorResponse`.\n`GET /map`, `GET /verifications/pending` and `GET /alerts` send an `ETag`; repeat the value in\n`If-None-Match` to get `304 Not Modified`.\n",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/command/shelters": {
            "get": {
                "tags": [
                    "Command"
                ],
                "summary": "All shelters (registry view, alphabetical)",
                "operationId": "get_api_command_shelter_list",
                "responses": {
                    "200": {
                        "description": "Shelters"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient role (error.code: forbidden)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Command"
                ],
                "summary": "Register a shelter",
                "operationId": "post_api_command_shelter_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateShelterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shelter created"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient role (error.code: forbidden)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/command/shelters/{id}": {
            "put": {
                "tags": [
                    "Command"
                ],
                "summary": "Update shelter master data (name, address, location, capacity)",
                "operationId": "put_api_command_shelter_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateShelterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated shelter"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient role (error.code: forbidden)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/command/shelters/{id}/status": {
            "post": {
                "tags": [
                    "Command"
                ],
                "summary": "Operator override of the shelter status (not counted as a citizen confirmation)",
                "operationId": "post_api_command_shelter_status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetShelterStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated shelter"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient role (error.code: forbidden)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/shelters": {
            "get": {
                "tags": [
                    "Shelters"
                ],
                "summary": "Shelters in a bbox, or the nearest ones to lat/lng",
                "operationId": "get_api_shelter_list",
                "parameters": [
                    {
                        "name": "bbox",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shelters (distanceMeters only with lat/lng)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ShelterView"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Malformed bbox (error.code: bad_request)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shelters/{id}": {
            "get": {
                "tags": [
                    "Shelters"
                ],
                "summary": "Shelter details",
                "operationId": "get_api_shelter_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shelter details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShelterView"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shelters/{id}/status": {
            "post": {
                "tags": [
                    "Shelters"
                ],
                "summary": "Confirm the current status of a shelter (open / closed / full)",
                "operationId": "post_api_shelter_confirm",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmShelterStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated shelter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShelterView"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorViolation": {
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "description": "Name of the body field (maps 1:1 onto the form field)",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorDetail": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "Stable machine-readable code: bad_request, unauthorized, token_expired, forbidden, not_found, conflict, gone, verification_already_answered, verification_expired, validation_failed, too_many_requests, internal_error.\n",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "retryAfter": {
                        "description": "Seconds until the next attempt may succeed (429 only, also sent as the Retry-After header)",
                        "type": "integer"
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorViolation"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                },
                "type": "object"
            },
            "GeoJsonGeometry": {
                "description": "GeoJSON geometry in WGS84, coordinates ordered [lng, lat].",
                "required": [
                    "type",
                    "coordinates"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "Point",
                            "Polygon",
                            "MultiPolygon"
                        ]
                    },
                    "coordinates": {
                        "description": "Point: [lng, lat]; Polygon: [[[lng, lat], ...]]; MultiPolygon: [[[[lng, lat], ...]]]",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "ShelterSummary": {
                "required": [
                    "id",
                    "name",
                    "status",
                    "statusLabel",
                    "confirmationCount"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "capacity": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/ShelterStatus"
                    },
                    "statusLabel": {
                        "type": "string"
                    },
                    "lastConfirmedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "confirmationCount": {
                        "type": "integer"
                    },
                    "distanceMeters": {
                        "description": "Only in GET /shelters?lat&lng",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ShelterView": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ShelterSummary"
                    },
                    {
                        "required": [
                            "location"
                        ],
                        "properties": {
                            "location": {
                                "$ref": "#/components/schemas/GeoJsonGeometry"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "CreateShelterRequest": {
                "required": [
                    "name",
                    "lat",
                    "lng"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 160
                    },
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "maximum": 90,
                        "minimum": -90
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "maximum": 180,
                        "minimum": -180
                    },
                    "address": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "maxLength": 255
                    },
                    "capacity": {
                        "type": "integer",
                        "nullable": true,
                        "default": null,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
            "UpdateShelterRequest": {
                "required": [
                    "name",
                    "lat",
                    "lng"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 160
                    },
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "maximum": 90,
                        "minimum": -90
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "maximum": 180,
                        "minimum": -180
                    },
                    "address": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "maxLength": 255
                    },
                    "capacity": {
                        "type": "integer",
                        "nullable": true,
                        "default": null,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
            "ShelterStatus": {
                "type": "string",
                "enum": [
                    "unknown",
                    "open",
                    "closed",
                    "full"
                ]
            },
            "SetShelterStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/ShelterStatus"
                    }
                },
                "type": "object"
            },
...
            "ConfirmShelterStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/ShelterStatus"
                    },
                    "comment": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "maxLength": 500
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}